Refactor both examples to reduce repeated code
