+++ examples/gencon.c
@@ -1,88 +1,7 @@
 /*
  * Generate connectivity (.co2) from Nek5000 mesh file (.re2).
  */
-#include <getopt.h>
-#include <mpi.h>
-#include <stdio.h>
-
 #include <parRSB.h>
-
-struct input {
-  char *mesh;
-  double tol;
-  int test;
-  int dump;
-  int nactive;
-};
-
-static int parse_input(struct input *in, int argc, char *argv[]) {
-  in->mesh = NULL;
-  in->tol = 0.2;
-  in->test = 0;
-  in->dump = 1;
-  MPI_Comm_size(MPI_COMM_WORLD, &in->nactive);
-
-  int c;
-  for (;;) {
-    static struct option long_options[] = {
-        {"mesh", required_argument, 0, 'm'},
-        {"tol", optional_argument, 0, 't'},
-        {"test", no_argument, 0, 'c'},
-        {"no-dump", no_argument, 0, 'd'},
-        {"nactive", optional_argument, 0, 'n'},
-        {0, 0, 0, 0}};
-
-    int option_index = 0;
-    c = getopt_long(argc, argv, "", long_options, &option_index);
-
-    if (c == -1)
-      break;
-
-    switch (c) {
-    case 'm':
-      in->mesh = optarg;
-      break;
-    case 't':
-      in->tol = atof(optarg);
-      break;
-    case 'c':
-      in->test = 1;
-      break;
-    case 'd':
-      in->dump = 0;
-      break;
-    case 'n':
-      in->nactive = atoi(optarg);
-      break;
-    case '?':
-      break;
-    default:
-      abort();
-    }
-  }
-}
-
-static void check_error_(int err, char *file, int line, MPI_Comm comm) {
-  int sum;
-  MPI_Allreduce(&err, &sum, 1, MPI_INT, MPI_SUM, comm);
-
-  if (sum != 0) {
-    int id;
-    MPI_Comm_rank(comm, &id);
-    if (id == 0)
-      printf("check_error failure in %s:%d\n", file, line);
-
-    MPI_Finalize();
-    exit(1);
-  }
-}
-
-#define check_error(err) check_error_(err, __FILE__, __LINE__, MPI_COMM_WORLD);
-
-struct pair {
-  ulong gid;
-  uint indx;
-};
 
 static int test_parcon(unsigned int neltp, long long *vlp, char *name,
                        MPI_Comm comm) {
@@ -160,34 +79,33 @@ static int test_parcon(unsigned int neltp, long long *vlp, char *name,
 int main(int argc, char *argv[]) {
   MPI_Init(&argc, &argv);
 
-  struct input in;
-  parse_input(&in, argc, argv);
+  parrsb_input *in = parrsb_parse_input(argc, argv);
 
   /* Read the geometry from the .re2 file */
   unsigned int nelt, nbcs;
   double *coord = NULL;
   long long *bcs = NULL;
   int nv;
-  int err = parrsb_read_mesh(&nelt, &nv, NULL, &coord, &nbcs, &bcs, in.mesh,
+  int err = parrsb_read_mesh(&nelt, &nv, NULL, &coord, &nbcs, &bcs, in->mesh,
                              MPI_COMM_WORLD, 1);
-  check_error(err);
+  parrsb_check_error(err, MPI_COMM_WORLD);
 
   /* Find connectivity */
   long long *vl = (long long *)calloc(nelt * nv, sizeof(long long));
   int ndim = nv == 8 ? 3 : 2;
-  err |= parRSB_findConnectivity(vl, coord, nelt, ndim, bcs, nbcs, in.tol,
+  err |= parRSB_findConnectivity(vl, coord, nelt, ndim, bcs, nbcs, in->tol,
                                  MPI_COMM_WORLD, 0);
-  check_error(err);
+  parrsb_check_error(err, MPI_COMM_WORLD);
 
   /* Test if the relevant env. variable is set */
-  if (in.test == 1)
-    err |= test_parcon(nelt, vl, in.mesh, MPI_COMM_WORLD);
-  check_error(err);
+  if (in->test == 1)
+    err |= test_parcon(nelt, vl, in->mesh, MPI_COMM_WORLD);
+  parrsb_check_error(err, MPI_COMM_WORLD);
 
   /* Write connectivity file */
-  if (in.dump == 1)
-    err |= parrsb_dump_con(vl, nelt, nv, in.mesh, MPI_COMM_WORLD);
-  check_error(err);
+  if (in->dump == 1)
+    err |= parrsb_dump_con(vl, nelt, nv, in->mesh, MPI_COMM_WORLD);
+  parrsb_check_error(err, MPI_COMM_WORLD);
 
   /* Free resources */
   if (vl != NULL)
@@ -196,10 +114,10 @@ int main(int argc, char *argv[]) {
     free(coord);
   if (bcs != NULL)
     free(bcs);
+  if (in != NULL)
+    free(in);
 
   MPI_Finalize();
 
   return 0;
 }
-
-#undef check_error
